Fix RBAC and add documentation

+++ src/main/java/com/preplane/dev/controllers/User/UserController.java
@@ -6,18 +6,19 @@ import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.preplane.dev.models.User;
 import com.preplane.dev.repositories.User.JDBCUserRepository;
+import com.preplane.dev.security.Auth;
 
 @CrossOrigin(origins = "http://localhost:8081")
 @RestController
@@ -27,9 +28,11 @@ public class UserController {
     @Autowired
     JDBCUserRepository userRepository;
 
-    @GetMapping("/user")
+    @GetMapping("/users")
+    @PreAuthorize("hasRole('ADMIN')")
     public ResponseEntity<List<User>> getAllUsers(@RequestParam Optional<Integer> limit,
             @RequestParam Optional<Integer> offset) {
+
         try {
             int lim = limit.orElse(50);
             int off = offset.orElse(0);
@@ -43,21 +46,17 @@ public class UserController {
 
     @GetMapping("/user/{id}")
     public ResponseEntity<User> getUserById(@PathVariable("id") int id) {
+
         var response = userRepository.findById(id);
         return new ResponseEntity<>(response.response, response.statusCode);
     }
 
-    @PutMapping("/user/{id}")
-    public ResponseEntity<String> updateUser(@PathVariable("id") int id,
-            @RequestBody User user) {
-
-        user.setId(id);
-        var response = userRepository.update(user);
-        return new ResponseEntity<>(response.message, response.statusCode);
-    }
-
     @DeleteMapping("/user/{id}")
     public ResponseEntity<String> deleteTutorial(@PathVariable("id") int id) {
+        // The user can only delete himself, or be deleted by an admin
+        if (Auth.getCurrentUser().getId() != id)
+            return new ResponseEntity<>(null, HttpStatus.UNAUTHORIZED);
+
         var response = userRepository.deleteById(id);
         return new ResponseEntity<>(response.message, response.statusCode);
     }
